Generate code using latest swagger spec

+++ RingCentral.Net/Definitions/BillingInfo.cs
@@ -5,11 +5,11 @@ namespace RingCentral
         /// <summary>
         /// Cost per minute, paid and already included in your RingCentral Plan. For example International Calls
         /// </summary>
-        public long? costIncluded;
+        public decimal? costIncluded;
 
         /// <summary>
         /// Cost per minute, paid and not included in your RingCentral Plan
         /// </summary>
-        public long? costPurchased;
+        public decimal? costPurchased;
     }
 }
